Changed directory structure; Split source file
